+++ app/src/main/java/ir/demoodite/dakhlokharj/data/room/models/DetailedPurchase.kt
@@ -3,7 +3,7 @@ package ir.demoodite.dakhlokharj.data.room.models
 data class DetailedPurchase(
     val purchaseId: Long,
     var purchaseProduct: String,
-    var purchasePrice: Long,
+    var purchasePrice: Double,
     var purchaseTime: Long,
     val purchaseBuyerId: Long,
     var buyerName: String,
