+++ app/src/main/java/ir/demoodite/dakhlokharj/data/room/models/Purchase.kt
@@ -24,7 +24,7 @@ import ir.demoodite.dakhlokharj.data.room.DataRepository.Companion.residentId
 data class Purchase(
     @PrimaryKey(autoGenerate = true) @ColumnInfo(name = purchaseId) val id: Long = 0,
     @ColumnInfo(name = purchaseProduct) var product: String = "",
-    @ColumnInfo(name = purchasePrice) var price: Long = 0,
+    @ColumnInfo(name = purchasePrice) var price: Double = 0.0,
     @ColumnInfo(name = purchaseBuyerId, index = true) var buyerId: Long = 0,
     @ColumnInfo(name = purchaseTime) var time: Long = 0,
 )
