+++ app/src/main/java/ir/demoodite/dakhlokharj/data/room/models/ResidentSummery.kt
@@ -2,6 +2,6 @@ package ir.demoodite.dakhlokharj.data.room.models
 
 data class ResidentSummery(
     val residentName: String,
-    val credit: Long,
-    val debt: Long,
+    val credit: Double,
+    val debt: Double,
 )
